Lift `selectedItems` getter up to `MainNavBar`
