+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbar.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useMemo, useState } from "react";
 import { connect } from "react-redux";
 import { push } from "react-router-redux";
 
@@ -8,8 +8,13 @@ import * as Urls from "metabase/lib/urls";
 import { closeNavbar, openNavbar } from "metabase/redux/app";
 
 import CollectionCreate from "metabase/collections/containers/CollectionCreate";
+import { coerceCollectionId } from "metabase/collections/utils";
 
-import { Collection } from "metabase-types/api";
+import { getQuestion } from "metabase/query_builder/selectors";
+import { getDashboard } from "metabase/dashboard/selectors";
+
+import Question from "metabase-lib/lib/Question";
+import { Collection, Dashboard } from "metabase-types/api";
 import { State } from "metabase-types/store";
 
 import DataAppNavbarContainer from "./DataAppNavbarContainer";
@@ -19,10 +24,25 @@ import {
   MainNavbarProps,
   MainNavbarOwnProps,
   MainNavbarDispatchProps,
+  SelectedItem,
 } from "./types";
 import { NavRoot, Sidebar } from "./MainNavbar.styled";
 
 type NavbarModal = "MODAL_NEW_COLLECTION" | null;
+
+interface StateProps {
+  question?: Question;
+  dashboard?: Dashboard;
+}
+
+type Props = MainNavbarProps & StateProps;
+
+function mapStateToProps(state: State) {
+  return {
+    question: getQuestion(state),
+    dashboard: getDashboard(state),
+  };
+}
 
 const mapDispatchToProps = {
   openNavbar,
@@ -33,13 +53,69 @@ const mapDispatchToProps = {
 function MainNavbar({
   isOpen,
   location,
+  params,
+  question,
+  dashboard,
   onChangeLocation,
   ...props
-}: MainNavbarProps) {
+}: Props) {
   const [modal, setModal] = useState<NavbarModal>(null);
 
   const isDataAppUrl = location.pathname.startsWith("/a/");
   const isDataAppPreview = location.pathname.startsWith("/a/preview/");
+
+  const selectedItems = useMemo<SelectedItem[]>(() => {
+    const { pathname } = location;
+    const { slug } = params;
+    const isCollectionPath = pathname.startsWith("/collection");
+    const isUsersCollectionPath = pathname.startsWith("/collection/users");
+    const isQuestionPath = pathname.startsWith("/question");
+    const isModelPath = pathname.startsWith("/model");
+    const isDataAppPath = pathname.startsWith("/a/");
+    const isDashboardPath = pathname.startsWith("/dashboard");
+
+    if (isCollectionPath) {
+      return [
+        {
+          id: isUsersCollectionPath ? "users" : Urls.extractCollectionId(slug),
+          type: "collection",
+        },
+      ];
+    }
+    if (isDataAppPath) {
+      return [
+        {
+          id: Urls.extractEntityId(slug),
+          type: "data-app",
+        },
+      ];
+    }
+    if (isDashboardPath && dashboard) {
+      return [
+        {
+          id: dashboard.id,
+          type: "dashboard",
+        },
+        {
+          id: coerceCollectionId(dashboard.collection_id),
+          type: "collection",
+        },
+      ];
+    }
+    if ((isQuestionPath || isModelPath) && question) {
+      return [
+        {
+          id: question.id(),
+          type: "card",
+        },
+        {
+          id: coerceCollectionId(question.collectionId()),
+          type: "collection",
+        },
+      ];
+    }
+    return [{ url: pathname, type: "non-entity" }];
+  }, [location, params, question, dashboard]);
 
   const onCreateNewCollection = useCallback(() => {
     setModal("MODAL_NEW_COLLECTION");
@@ -70,6 +146,8 @@ function MainNavbar({
             <DataAppNavbarContainer
               isOpen={isOpen}
               location={location}
+              params={params}
+              selectedItems={selectedItems}
               onChangeLocation={onChangeLocation}
               {...props}
             />
@@ -77,6 +155,8 @@ function MainNavbar({
             <MainNavbarContainer
               isOpen={isOpen}
               location={location}
+              params={params}
+              selectedItems={selectedItems}
               onCreateNewCollection={onCreateNewCollection}
               onChangeLocation={onChangeLocation}
               {...props}
@@ -90,11 +170,11 @@ function MainNavbar({
 }
 
 export default connect<
-  unknown,
+  StateProps,
   MainNavbarDispatchProps,
   MainNavbarOwnProps,
   State
 >(
-  null,
+  mapStateToProps,
   mapDispatchToProps,
 )(MainNavbar);
